Calculate the quadratic Fleiss' kappa and Conger's (Cohen's) kappa

Estimate a quadratically weighted kappa with potentially missing data using least squares, or calculate its population value. conger and cohen are aliases. See quadagree() and cohenci() for confidence intervals.

Usage

fleiss(x)

conger(x)

cohen(x)

conger_pop(mu, sigma)

cohen_pop(mu, sigma)

fleiss_pop(mu, sigma)

fleiss_aggr(x, values = seq(ncol(x)))

Arguments

x

Input data on wide form. Columns are raters, items are ratings. Missing values are allowed, and should be encoded as NAs.

mu

Population vector of means.

sigma

Population covariance matrix.

values

to attach to each column on the Fleiss form data. Defaults to 1:C, where C is the number of categories. Only used in fleiss_aggr.

Value

Sample / population quadratically weighted Fleiss' or Conger's kappa.

Details

Conger's kappa is a multi-rater generalization of Cohen's kappa. All functions in this package work for multiple raters, so functions starting with cohen or conger are aliases. The quadratically weighted Cohen's kappa is also known as Lin's (1989) concordance coefficient.

The only difference between Cohen's kappa and Fleiss' kappa lies on how they measure disagreement due to chance. Here Fleiss' marginalizes the rating distribution across raters, essentially assuming there is no difference in the rating distribution across raters, while Cohen's kappa does not. There is a large literature comparing Fleiss' kappa to Cohen's kappa, and there is no consensus on which to prefer.

The functions cohen_pop, conger_pop, and fleiss_pop calculate the population values of the kappas using the mean and covariance of the rating distributions. The derivation of the covariance and mean formulation can be found in (Moss and van Oest, work in progress.)

The functions cohen, conger, and fleiss estimate the kappas using least squares. This method uses the biased sample covariance estimator, which agrees with the literature on agreement coefficients in the case of two raters. In the case of missing data, the pairwise available data is used, employing the option use = "pairwise.complete.obs" in stats::cov().

References

Cohen, J. (1968). Weighted kappa: Nominal scale agreement with provision for scaled disagreement or partial credit. Psychological Bulletin, 70(4), 213–220. https://doi.org/10.1037/h0026256

Fleiss, J. L. (1975). Measuring agreement between two judges on the presence or absence of a trait. Biometrics, 31(3), 651–659. https://www.ncbi.nlm.nih.gov/pubmed/1174623

Conger, A. J. (1980). Integration and generalization of kappas for multiple raters. Psychological Bulletin, 88(2), 322–328. https://doi.org/10.1037/0033-2909.88.2.322

Lin, L. I. (1989). A concordance correlation coefficient to evaluate reproducibility. Biometrics, 45(1), 255–268. https://www.ncbi.nlm.nih.gov/pubmed/2720055

Moss, van Oest (work in progress). Inference for quadratically weighted multi-rater kappas with missing raters.

Examples

x <- irrCAC::cac.raw4raters[2:9, ]
fleiss(x)
#> [1] 0.6666667
# [1] 0.6666667
irrCAC::fleiss.kappa.raw(x, weights = "quadratic")$est$coeff.val
#> [1] 0.66667
# [1] 0.66667 (irrCAC prematurely rounds the coefficient)

x <- irrCAC::cac.raw4raters[2:9, ]
conger(x)
#> [1] 0.6719243
# [1] 0.6719243
irrCAC::conger.kappa.raw(x, weights = "quadratic")$est$coeff.val
#> [1] 0.67192
# [1] 0.67192 (irrCAC prematurely rounds the coefficient)